requirements.txt
tree-sitter==0.19.0
tree-sitter-python
tree-sitter-languages
py-tree-sitter
numpy==1.24.0
openai==0.27.0
tensorflow-hub==0.13.0
fastapi==0.95.2
python-dotenv
scikit-learn==1.2.2
fitz==0.0.1.dev2  
gradio==3.39.0
requests==2.31.0
openai
gradio
